vim: Replace plug with packages/pathogen
I decided to ditch vim-plug because I realised that I could replicate
the parts of it that I used very easily. I decided to use packages when
it was available and pathogen when it wasn't because pathogen seems to
have some edge cases in which it does not work as expected, while
packages is built into vim.