pack: Fix vim-vint/libyaml compat and update all
Basically, pyyaml bugs out when you install it if libyaml is installed.
This commit removes libyaml before installing pyyaml, and adds it back
again after it is done. This commit also adds sections to this growing
script, and also makes sure to update pip and ruby packages (this gets
a little tricky with the whole libyaml thing).