bash: Partly disable bash-completion
I did this because the vast majority of the commands that
bash-completion provides completions for I use very rarely. In addition
to this, the only commands I use on a frequent basis that have 'complex'
structures are git and brew, which is why I still source the completions
for those.